+++ utils.go
@@ -1,0 +1,16 @@
+package owl
+
+import (
+	"os"
+)
+
+func exists(path string) (bool, error) {
+	stat, err := os.Stat(path)
+	if err == nil {
+		return !stat.IsDir(), err
+	}
+	if os.IsNotExist(err) {
+		return false, nil
+	}
+	return false, err
+}